tree2: Split editable-tree-2 into flex-tree and simple-tree (#18294)
## Description

Refactor editable-tree-2 into flex-tree and simple-tree.

The tests for "proxy" were never organized into the folder hierarchy to
match the source files, so they were a bit tangled with the
editable-tree test files: this was sorted out mostly cleanly, though a
utils file had to be split, and some files (like the events tests)
likely should be split later as well.

`LazyTree` was renamed to `LazyNode`.

This change will be followed by some major renames:
1. The types in flex tree will get prefixed with Flex.
2. The types in simple-tree with use the generic names that flex tree
used to use.
