+++ experimental/dds/tree2/.dependency-cruiser-known-violations.json
@@ -25,23 +25,23 @@
 	},
 	{
 		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/context.ts",
-		"to": "src/feature-libraries/editable-tree-2/lazyField.ts",
+		"from": "src/feature-libraries/flex-tree/context.ts",
+		"to": "src/feature-libraries/flex-tree/lazyField.ts",
 		"rule": {
 			"severity": "error",
 			"name": "no-circular"
 		},
 		"cycle": [
-			"src/feature-libraries/editable-tree-2/lazyField.ts",
+			"src/feature-libraries/flex-tree/lazyField.ts",
 			"src/feature-libraries/node-key/index.ts",
 			"src/feature-libraries/node-key/nodeKeyIndex.ts",
-			"src/feature-libraries/editable-tree-2/index.ts",
-			"src/feature-libraries/editable-tree-2/context.ts"
+			"src/feature-libraries/flex-tree/index.ts",
+			"src/feature-libraries/flex-tree/context.ts"
 		]
 	},
 	{
 		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/context.ts",
+		"from": "src/feature-libraries/flex-tree/context.ts",
 		"to": "src/feature-libraries/node-key/index.ts",
 		"rule": {
 			"severity": "error",
@@ -50,134 +50,47 @@
 		"cycle": [
 			"src/feature-libraries/node-key/index.ts",
 			"src/feature-libraries/node-key/nodeKeyIndex.ts",
-			"src/feature-libraries/editable-tree-2/index.ts",
-			"src/feature-libraries/editable-tree-2/context.ts"
+			"src/feature-libraries/flex-tree/index.ts",
+			"src/feature-libraries/flex-tree/context.ts"
 		]
 	},
 	{
 		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/index.ts",
-		"to": "src/feature-libraries/editable-tree-2/proxies/index.ts",
+		"from": "src/feature-libraries/flex-tree/lazyField.ts",
+		"to": "src/feature-libraries/flex-tree/lazyNode.ts",
 		"rule": {
 			"severity": "error",
 			"name": "no-circular"
 		},
 		"cycle": [
-			"src/feature-libraries/editable-tree-2/proxies/index.ts",
-			"src/feature-libraries/editable-tree-2/proxies/node.ts",
-			"src/feature-libraries/editable-tree-2/proxies/proxies.ts",
-			"src/feature-libraries/editable-tree-2/lazyTree.ts",
-			"src/feature-libraries/editable-tree-2/lazyField.ts",
-			"src/feature-libraries/node-key/index.ts",
-			"src/feature-libraries/node-key/nodeKeyIndex.ts",
-			"src/feature-libraries/editable-tree-2/index.ts"
+			"src/feature-libraries/flex-tree/lazyNode.ts",
+			"src/feature-libraries/flex-tree/lazyField.ts"
 		]
 	},
 	{
 		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/lazyField.ts",
-		"to": "src/feature-libraries/editable-tree-2/lazyTree.ts",
+		"from": "src/feature-libraries/flex-tree/lazyField.ts",
+		"to": "src/feature-libraries/flex-tree/unboxed.ts",
 		"rule": {
 			"severity": "error",
 			"name": "no-circular"
 		},
 		"cycle": [
-			"src/feature-libraries/editable-tree-2/lazyTree.ts",
-			"src/feature-libraries/editable-tree-2/lazyField.ts"
+			"src/feature-libraries/flex-tree/unboxed.ts",
+			"src/feature-libraries/flex-tree/lazyField.ts"
 		]
 	},
 	{
 		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/lazyField.ts",
-		"to": "src/feature-libraries/editable-tree-2/unboxed.ts",
+		"from": "src/feature-libraries/flex-tree/lazyNode.ts",
+		"to": "src/feature-libraries/flex-tree/unboxed.ts",
 		"rule": {
 			"severity": "error",
 			"name": "no-circular"
 		},
 		"cycle": [
-			"src/feature-libraries/editable-tree-2/unboxed.ts",
-			"src/feature-libraries/editable-tree-2/lazyField.ts"
-		]
-	},
-	{
-		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/lazyField.ts",
-		"to": "src/feature-libraries/editable-tree-2/utilities.ts",
-		"rule": {
-			"severity": "error",
-			"name": "no-circular"
-		},
-		"cycle": [
-			"src/feature-libraries/editable-tree-2/utilities.ts",
-			"src/feature-libraries/editable-tree-2/index.ts",
-			"src/feature-libraries/editable-tree-2/context.ts",
-			"src/feature-libraries/editable-tree-2/lazyField.ts"
-		]
-	},
-	{
-		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/lazyTree.ts",
-		"to": "src/feature-libraries/editable-tree-2/unboxed.ts",
-		"rule": {
-			"severity": "error",
-			"name": "no-circular"
-		},
-		"cycle": [
-			"src/feature-libraries/editable-tree-2/unboxed.ts",
-			"src/feature-libraries/editable-tree-2/lazyTree.ts"
-		]
-	},
-	{
-		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/lazyTree.ts",
-		"to": "src/feature-libraries/editable-tree-2/utilities.ts",
-		"rule": {
-			"severity": "error",
-			"name": "no-circular"
-		},
-		"cycle": [
-			"src/feature-libraries/editable-tree-2/utilities.ts",
-			"src/feature-libraries/editable-tree-2/index.ts",
-			"src/feature-libraries/editable-tree-2/context.ts",
-			"src/feature-libraries/editable-tree-2/lazyField.ts",
-			"src/feature-libraries/editable-tree-2/lazyTree.ts"
-		]
-	},
-	{
-		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/proxies/index.ts",
-		"to": "src/feature-libraries/editable-tree-2/proxies/objectFactory.ts",
-		"rule": {
-			"severity": "error",
-			"name": "no-circular"
-		},
-		"cycle": [
-			"src/feature-libraries/editable-tree-2/proxies/objectFactory.ts",
-			"src/feature-libraries/editable-tree-2/proxies/proxies.ts",
-			"src/feature-libraries/editable-tree-2/lazyTree.ts",
-			"src/feature-libraries/editable-tree-2/lazyField.ts",
-			"src/feature-libraries/node-key/index.ts",
-			"src/feature-libraries/node-key/nodeKeyIndex.ts",
-			"src/feature-libraries/editable-tree-2/index.ts",
-			"src/feature-libraries/editable-tree-2/proxies/index.ts"
-		]
-	},
-	{
-		"type": "cycle",
-		"from": "src/feature-libraries/editable-tree-2/proxies/index.ts",
-		"to": "src/feature-libraries/editable-tree-2/proxies/proxies.ts",
-		"rule": {
-			"severity": "error",
-			"name": "no-circular"
-		},
-		"cycle": [
-			"src/feature-libraries/editable-tree-2/proxies/proxies.ts",
-			"src/feature-libraries/editable-tree-2/lazyTree.ts",
-			"src/feature-libraries/editable-tree-2/lazyField.ts",
-			"src/feature-libraries/node-key/index.ts",
-			"src/feature-libraries/node-key/nodeKeyIndex.ts",
-			"src/feature-libraries/editable-tree-2/index.ts",
-			"src/feature-libraries/editable-tree-2/proxies/index.ts"
+			"src/feature-libraries/flex-tree/unboxed.ts",
+			"src/feature-libraries/flex-tree/lazyNode.ts"
 		]
 	},
 	{
