+++ experimental/dds/tree2/src/feature-libraries/flex-tree/lazyNode.ts
@@ -36,9 +36,9 @@ import {
 	Any,
 	AllowedTypes,
 } from "../typed-schema";
-import { EditableTreeEvents, TreeEvent } from "../untypedTree";
 import { FieldKinds } from "../default-field-kinds";
 import { LocalNodeKey } from "../node-key";
+import { EditableTreeEvents, TreeEvent } from "./treeEvents";
 import { Context } from "./context";
 import {
 	FieldNode,
