+++ experimental/dds/tree2/src/feature-libraries/index.ts
@@ -215,8 +215,6 @@ export {
 	defaultSchemaPolicy,
 } from "./default-field-kinds";
 
-export { TreeEvent, EditableTreeEvents } from "./untypedTree";
-
 export {
 	AssignableFieldKinds,
 	FieldNode,
@@ -242,6 +240,13 @@ export {
 	boxedIterator,
 	CheckTypesOverlap,
 	TreeStatus,
+	Typed,
+	Context,
+	TreeEvent,
+	EditableTreeEvents,
+} from "./flex-tree";
+
+export {
 	getProxyForField,
 	ObjectFields,
 	ProxyField,
@@ -255,12 +260,10 @@ export {
 	Tree,
 	TreeApi,
 	SharedTreeNode,
-	Typed,
 	SharedTreeObjectFactory,
 	FactoryTreeSchema,
 	addFactory,
-	Context,
-} from "./editable-tree-2";
+} from "./simple-tree";
 
 export { treeSchemaFromStoredSchema } from "./storedToViewSchema";
