+++ experimental/dds/tree2/src/feature-libraries/simple-tree/README.md
@@ -1,0 +1,3 @@
+# simple-tree
+
+Simple-Tree is a simple to use wrapper around flex-tree.
