+++ experimental/dds/tree2/src/feature-libraries/simple-tree/objectFactory.ts
@@ -3,8 +3,8 @@
  * Licensed under the MIT License.
  */
 
-import { Assume } from "../../../util";
-import { ObjectNodeSchema, TreeNodeSchemaBase } from "../../typed-schema";
+import { Assume } from "../../util";
+import { ObjectNodeSchema, TreeNodeSchemaBase } from "../typed-schema";
 import { createRawObjectProxy } from "./proxies";
 import { ProxyNode, SharedTreeObject } from "./types";
