+++ experimental/dds/tree2/src/test/feature-libraries/flex-tree/utils.ts
@@ -6,19 +6,14 @@
 import { strict as assert } from "node:assert";
 import {
 	DefaultEditBuilder,
-	TreeFieldSchema,
-	ImplicitFieldSchema,
-	ProxyField,
-	ProxyRoot,
 	TreeSchema,
 	createMockNodeKeyManager,
 	nodeKeyFieldKey,
-	SchemaAware,
 	AllowedTypes,
 	FieldKind,
 } from "../../../feature-libraries";
 // eslint-disable-next-line import/no-internal-modules
-import { Context, getTreeContext } from "../../../feature-libraries/editable-tree-2/context";
+import { Context, getTreeContext } from "../../../feature-libraries/flex-tree/context";
 import {
 	FieldAnchor,
 	IEditableForest,
@@ -27,9 +22,8 @@ import {
 	rootFieldKey,
 } from "../../../core";
 import { TreeContent } from "../../../shared-tree";
-import { forestWithContent, viewWithContent } from "../../utils";
+import { forestWithContent } from "../../utils";
 import { brand } from "../../../util";
-import { SchemaBuilder } from "../../../domains";
 
 export function getReadonlyContext(forest: IEditableForest, schema: TreeSchema): Context {
 	// This will error if someone tries to call mutation methods on it
@@ -84,56 +78,4 @@ export function readonlyTreeWithContent<Kind extends FieldKind, Types extends Al
 		context,
 		cursor,
 	};
-}
-
-/** Helper for making small test schemas. */
-export function makeSchema<const TSchema extends ImplicitFieldSchema>(
-	fn: (builder: SchemaBuilder) => TSchema,
-) {
-	const builder = new SchemaBuilder({
-		scope: `test.schema.${Math.random().toString(36).slice(2)}`,
-	});
-	const root = fn(builder);
-	return builder.intoSchema(root);
-}
-
-/**
- * @deprecated Write a normal `it` test. Doing so allows:
- * 1. Selecting between viewWithContent and {@link readonlyTreeWithContent} or some other setup.
- * 2. Navigate to test source and similar tools and IDE integration to work.
- * 3. Use of `it.only` and `it.skip`.
- * 4. Easier understanding of what is a test for new people looking at the test files.
- * 5. Consistent test patterns for users of APIs other than context.root.
- * 6. Ability to write async tests.
- */
-export function itWithRoot<TRoot extends TreeFieldSchema>(
-	title: string,
-	schema: TreeSchema<TRoot>,
-	initialTree: ProxyRoot<TreeSchema<TRoot>, "javaScript">,
-	fn: (root: ProxyField<(typeof schema)["rootFieldSchema"]>) => void,
-): void {
-	it(title, () => {
-		const view = viewWithContent({
-			schema,
-			initialTree: initialTree as SchemaAware.TypedField<TRoot, SchemaAware.ApiMode.Flexible>,
-		});
-		const root = view.root;
-		fn(root);
-	});
-}
-
-/**
- * Similar to JSON stringify, but preserves `undefined` and numbers numbers as-is at the root.
- */
-export function pretty(arg: unknown): number | undefined | string {
-	if (arg === undefined) {
-		return "undefined";
-	}
-	if (typeof arg === "number") {
-		return arg;
-	}
-	if (typeof arg === "string") {
-		return `"${arg}"`;
-	}
-	return JSON.stringify(arg);
 }
